requirements.txt
allure_python_commons==2.15.3
pytest==8.4.2
Requests==2.33.1
